+++ internal/acctest/configs.go
@@ -295,6 +295,17 @@ provider "aws" {
 `, tag1, value1, tag2, value2))
 }
 
+func ConfigAssumeRole() string {
+	//lintignore:AT004
+	return fmt.Sprintf(`
+provider "aws" {
+  assume_role {
+    role_arn = %[1]q
+  }
+}
+`, os.Getenv(envvar.AccAssumeRoleARN))
+}
+
 func ConfigAssumeRolePolicy(policy string) string {
 	//lintignore:AT004
 	return fmt.Sprintf(`
@@ -606,7 +617,7 @@ resource "aws_security_group" "sg_for_lambda" {
 
 func ConfigVPCWithSubnets(rName string, subnetCount int) string {
 	return ConfigCompose(
-		ConfigAvailableAZsNoOptInDefaultExclude(),
+		ConfigSubnets(rName, subnetCount),
 		fmt.Sprintf(`
 resource "aws_vpc" "test" {
   cidr_block = "10.0.0.0/16"
@@ -615,19 +626,7 @@ resource "aws_vpc" "test" {
     Name = %[1]q
   }
 }
-
-resource "aws_subnet" "test" {
-  count = %[2]d
-
-  vpc_id            = aws_vpc.test.id
-  availability_zone = data.aws_availability_zones.available.names[count.index]
-  cidr_block        = cidrsubnet(aws_vpc.test.cidr_block, 8, count.index)
-
-  tags = {
-    Name = %[1]q
-  }
-}
-`, rName, subnetCount),
+`, rName),
 	)
 }
 
@@ -656,7 +655,7 @@ resource "aws_subnet" "test" {
 
 func ConfigVPCWithSubnetsEnableDNSHostnames(rName string, subnetCount int) string {
 	return ConfigCompose(
-		ConfigAvailableAZsNoOptInDefaultExclude(),
+		ConfigSubnets(rName, subnetCount),
 		fmt.Sprintf(`
 resource "aws_vpc" "test" {
   cidr_block           = "10.0.0.0/16"
@@ -666,7 +665,31 @@ resource "aws_vpc" "test" {
     Name = %[1]q
   }
 }
+`, rName),
+	)
+}
 
+func ConfigVPCWithSubnetsIPv6(rName string, subnetCount int) string {
+	return ConfigCompose(
+		ConfigSubnetsIPv6(rName, subnetCount),
+		fmt.Sprintf(`
+resource "aws_vpc" "test" {
+  cidr_block = "10.0.0.0/16"
+
+  assign_generated_ipv6_cidr_block = true
+
+  tags = {
+    Name = %[1]q
+  }
+}
+`, rName),
+	)
+}
+
+func ConfigSubnets(rName string, subnetCount int) string {
+	return ConfigCompose(
+		ConfigAvailableAZsNoOptInDefaultExclude(),
+		fmt.Sprintf(`
 resource "aws_subnet" "test" {
   count = %[2]d
 
@@ -682,20 +705,10 @@ resource "aws_subnet" "test" {
 	)
 }
 
-func ConfigVPCWithSubnetsIPv6(rName string, subnetCount int) string {
+func ConfigSubnetsIPv6(rName string, subnetCount int) string {
 	return ConfigCompose(
 		ConfigAvailableAZsNoOptInDefaultExclude(),
 		fmt.Sprintf(`
-resource "aws_vpc" "test" {
-  cidr_block = "10.0.0.0/16"
-
-  assign_generated_ipv6_cidr_block = true
-
-  tags = {
-    Name = %[1]q
-  }
-}
-
 resource "aws_subnet" "test" {
   count = %[2]d